feat: add more output formats
Support an explicit `format` option and new functionality for the format values of `element` and
`fragment`. Also support the default module id of `jquery` when supporting the `jquery` format.

BREAKING CHANGE: mostly backwards compatible, but some edge cases related to formats that might
crop up; it's best to just update the template invocations to use the format option intentionally.